Refactor @mahler/testing sub-module
This breaks up index into multiple files to simplify the move to
a DAG representation of a plan.

Change-type: minor
