+++ lib/testing/types.ts
@@ -1,0 +1,7 @@
+/**
+ * A "simplified" plan.
+ *
+ * It is an object representation
+ * of a plan that's easier to print and compare.
+ */
+export type SimplePlan = string[];
